% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARPH2-package.R
\docType{package}
\name{PAWPH2-package}
\alias{PAWPH2}
\alias{PAWPH2-package}
\title{PAWPH2: Penalized Adaptive Robust Proportional Hazard Model}
\description{
Fits the Penalized Adaptive Weighted Proportional Hazard (PAWPH) model for robust survival analysis in the presence of outliers. Uses per-observation intercepts and an adaptive lasso penalty to simultaneously perform variable selection and outlier detection. Incorporates a modified Cox regularisation routine adapted from the ncvreg package (Breheny and Huang, 2011, \doi{10.1214/10-AOAS388}).
}
\author{
\strong{Maintainer}: Garth Tarr \email{garth.tarr@sydney.edu.au}

Other contributors:
\itemize{
  \item Patrick Breheny \email{patrick-breheny@uiowa.edu} (\href{https://orcid.org/0000-0002-0650-1119}{ORCID}) (Author of ncvreg; C and R code adapted from ncvreg v3.13.0) [contributor]
}

}
\keyword{internal}
